+++ backend/ResellioBackend/.idea/.idea.ResellioBackend/.idea/workspace.xml
@@ -9,10 +9,21 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="97154ca5-6890-4747-ac0c-399e362971eb" name="Changes" comment="">
-      <change afterPath="$PROJECT_DIR$/UserManagementSystem/Services/Implementations/EventService.cs" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/UserManagementSystem/Services/Implementations/TicketService.cs" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/UserManagementSystem/Services/Implementations/TicketTypeService.cs" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/EventManagementSystem/Services/Abstractions/IEventService.cs" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/EventManagementSystem/Services/Implementations/EventService.cs" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/.idea.ResellioBackend/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/.idea.ResellioBackend/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/UserManagementSystem/DTOs/EventDTO.cs" beforeDir="false" afterPath="$PROJECT_DIR$/EventManagementSystem/DTOs/EventDTO.cs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/UserManagementSystem/DTOs/TicketDTO.cs" beforeDir="false" afterPath="$PROJECT_DIR$/EventManagementSystem/DTOs/TicketDTO.cs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/UserManagementSystem/DTOs/TicketTypeDTO.cs" beforeDir="false" afterPath="$PROJECT_DIR$/EventManagementSystem/DTOs/TicketTypeDTO.cs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/UserManagementSystem/Models/Event.cs" beforeDir="false" afterPath="$PROJECT_DIR$/EventManagementSystem/Models/Event.cs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/UserManagementSystem/Models/Ticket.cs" beforeDir="false" afterPath="$PROJECT_DIR$/EventManagementSystem/Models/Ticket.cs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/UserManagementSystem/Models/TicketType.cs" beforeDir="false" afterPath="$PROJECT_DIR$/EventManagementSystem/Models/TicketType.cs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/UserManagementSystem/Services/Abstractions/IEventService.cs" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/UserManagementSystem/Services/Abstractions/ITicketService.cs" beforeDir="false" afterPath="$PROJECT_DIR$/EventManagementSystem/Services/Abstractions/ITicketService.cs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/UserManagementSystem/Services/Abstractions/ITicketTypeService.cs" beforeDir="false" afterPath="$PROJECT_DIR$/EventManagementSystem/Services/Abstractions/ITicketTypeService.cs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/UserManagementSystem/Services/Implementations/EventService.cs" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/UserManagementSystem/Services/Implementations/TicketService.cs" beforeDir="false" afterPath="$PROJECT_DIR$/EventManagementSystem/Services/Implementations/TicketService.cs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/UserManagementSystem/Services/Implementations/TicketTypeService.cs" beforeDir="false" afterPath="$PROJECT_DIR$/EventManagementSystem/Services/Implementations/TicketTypeService.cs" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -22,33 +33,33 @@
   <component name="Git.Settings">
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$/../.." />
   </component>
-  <component name="ProjectColorInfo"><![CDATA[{
-  "associatedIndex": 3
-}]]></component>
+  <component name="ProjectColorInfo">{
+  &quot;associatedIndex&quot;: 3
+}</component>
   <component name="ProjectId" id="2ttlTPvXKEmEOyekj99rxT6SHeT" />
   <component name="ProjectViewState">
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
-  <component name="PropertiesComponent"><![CDATA[{
-  "keyToString": {
-    "RunOnceActivity.OpenProjectViewOnStart": "true",
-    "RunOnceActivity.ShowReadmeOnStart": "true",
-    "git-widget-placeholder": "feat/event-creation",
-    "node.js.detected.package.eslint": "true",
-    "node.js.selected.package.eslint": "(autodetect)",
-    "node.js.selected.package.tslint": "(autodetect)",
-    "nodejs_package_manager_path": "npm",
-    "vue.rearranger.settings.migration": "true"
+  <component name="PropertiesComponent">{
+  &quot;keyToString&quot;: {
+    &quot;RunOnceActivity.OpenProjectViewOnStart&quot;: &quot;true&quot;,
+    &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
+    &quot;git-widget-placeholder&quot;: &quot;feat/event-creation&quot;,
+    &quot;node.js.detected.package.eslint&quot;: &quot;true&quot;,
+    &quot;node.js.selected.package.eslint&quot;: &quot;(autodetect)&quot;,
+    &quot;node.js.selected.package.tslint&quot;: &quot;(autodetect)&quot;,
+    &quot;nodejs_package_manager_path&quot;: &quot;npm&quot;,
+    &quot;vue.rearranger.settings.migration&quot;: &quot;true&quot;
   },
-  "keyToStringList": {
-    "rider.external.source.directories": [
-      "/Users/miloszwysocki/Library/Application Support/JetBrains/Rider2023.3/resharper-host/DecompilerCache",
-      "/Users/miloszwysocki/Library/Application Support/JetBrains/Rider2023.3/resharper-host/SourcesCache",
-      "/Users/miloszwysocki/.local/share/Symbols/src"
+  &quot;keyToStringList&quot;: {
+    &quot;rider.external.source.directories&quot;: [
+      &quot;/Users/miloszwysocki/Library/Application Support/JetBrains/Rider2023.3/resharper-host/DecompilerCache&quot;,
+      &quot;/Users/miloszwysocki/Library/Application Support/JetBrains/Rider2023.3/resharper-host/SourcesCache&quot;,
+      &quot;/Users/miloszwysocki/.local/share/Symbols/src&quot;
     ]
   }
-}]]></component>
+}</component>
   <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
   <component name="TaskManager">
     <task active="true" id="Default" summary="Default task">
@@ -57,7 +68,8 @@
       <option name="number" value="Default" />
       <option name="presentableId" value="Default" />
       <updated>1741179259997</updated>
-      <workItem from="1741179261797" duration="6370000" />
+      <workItem from="1741179261797" duration="6876000" />
+      <workItem from="1741186747194" duration="1072000" />
     </task>
     <servers />
   </component>
