+++ backend/ResellioBackend/EventManagementSystem/Services/Abstractions/IEventService.cs
@@ -1,0 +1,9 @@
+using ResellioBackend.EventManagementSystem.DTOs;
+using ResellioBackend.UserManagementSystem.Models;
+
+namespace ResellioBackend.EventManagementSystem.Services.Abstractions;
+
+public interface IEventService
+{
+    Task<Event> CreateEventAsync(EventDTO eventDto, int organiserId);
+}
